# health-tracker-docs/yuanqi/health-tracker-openapi.yaml
openapi: 3.0.0
info:
  title: 健康管理小程序 API
  version: 1.0.0
  description: 供腾讯元器插件导入的 OpenAPI 文档（简化版）
servers:
  - url: "http://139.196.155.234:8080"
    description: "生产环境 API 服务器"
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    ApiResponse:
      type: object
      properties:
        code:
          type: integer
          example: 0
        message:
          type: string
          example: ok
        data:
          type: object
    ReminderAddRequest:
      type: object
      required: [userId, title, type, remindTime]
      properties:
        userId:
          type: integer
        title:
          type: string
        type:
          type: integer
          description: "1=运动,2=饮食,3=睡眠"
        content:
          type: string
        remindTime:
          type: string
          example: "2026-03-15 20:30:00"
...
    MedicationAddRequest:
      type: object
      required: [userId, drugName, dosage, frequency]
      properties:
        userId:
          type: integer
        drugName:
          type: string
        dosage:
          type: string
        frequency:
          type: string
        remindTime:
          type: string
          example: "08:00"
        startDate:
          type: string
          example: "2026-03-15"
...
security:
  - bearerAuth: []
paths:
...
  /api/user/profile:
    get:
      summary: 获取用户资料
      parameters:
        - in: query
          name: userId
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: 成功
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponse"
  /api/user/profile/update:
    post:
      summary: 更新用户资料
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [userId]
              properties:
                userId:
                  type: integer
                wxNickname:
                  type: string
                wxAvatar:
                  type: string
                sex:
                  type: string
                age:
                  type: integer
                height:
                  type: number
                weight:
                  type: number
                systolic:
                  type: integer
                diastolic:
                  type: integer
                heartRate:
                  type: integer
      responses:
        "200":
          description: 更新成功
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponse"
...
  /api/reminder/list:
    get:
      summary: 提醒列表
      parameters:
        - in: query
          name: userId
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: 成功
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponse"
  /api/reminder/add:
    post:
      summary: 新增提醒
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ReminderAddRequest"
      responses:
        "200":
          description: 成功
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponse"
  /api/reminder/update:
    post:
      summary: 更新提醒
      requestBody:
        required: true
        content:
          application/json:
            schema:
              allOf:
                - $ref: "#/components/schemas/ReminderAddRequest"
                - type: object
                  required: [id]
                  properties:
                    id:
                      type: integer
      responses:
        "200":
          description: 成功
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponse"
...
  /api/medication/list:
    get:
      summary: 用药列表
      parameters:
        - in: query
          name: userId
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: 成功
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponse"
  /api/medication/add:
    post:
      summary: 新增用药
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MedicationAddRequest"
      responses:
        "200":
          description: 成功
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponse"
  /api/medication/update:
    post:
      summary: 更新用药
      requestBody:
        required: true
        content:
          application/json:
            schema:
              allOf:
                - $ref: "#/components/schemas/MedicationAddRequest"
                - type: object
                  required: [id]
                  properties:
                    id:
                      type: integer
      responses:
        "200":
          description: 成功
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponse"
...